requirements.txt
numpy
pandas==1.5.3
bw2io==0.8.7
bw2data==3.6.6
wurst==0.3.3
xarray
prettytable
pycountry
cryptography
premise_gwp
pyYaml
sparse
schema
datapackage
requests
bottleneck
constructive_geometries>=0.8.2
